+++ src/main/java/tosiltosil/backend/module/goal/application/GoalService.java
@@ -17,6 +17,7 @@ import tosiltosil.backend.module.goal.domain.request.GoalCreateRequest;
 import tosiltosil.backend.module.goal.domain.request.GoalOrderChangeRequest;
 import tosiltosil.backend.module.goal.domain.request.GoalUpdateRequest;
 import tosiltosil.backend.module.goal.domain.response.DayGoalsResponse;
+import tosiltosil.backend.module.goal.domain.response.GoalListPerCategoryResponse;
 import tosiltosil.backend.module.goal.domain.response.GoalIdResponse;
 import tosiltosil.backend.module.goal.domain.response.GoalIdsResponse;
 import tosiltosil.backend.module.goal.domain.response.GoalOrderChangeResponse;
@@ -38,7 +39,7 @@ public class GoalService {
     ) {
         // TODO: 친구 관계 검증
         
-        List<DayGoalsResponse.GoalListPerCategoryResponse> categoryResponses = goalRepository.findDayGoals(memberId, date);
+        List<GoalListPerCategoryResponse> categoryResponses = goalRepository.findDayGoals(memberId, date);
 
         BigDecimal overallPercentage = goalDomainService.calculateGoalAchievedPercentage(memberId);
