+++ src/main/java/tosiltosil/backend/module/goal/domain/GoalRepository.java
@@ -5,13 +5,13 @@ import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
-import tosiltosil.backend.module.goal.domain.response.DayGoalsResponse;
+import tosiltosil.backend.module.goal.domain.response.GoalListPerCategoryResponse;
 
 public interface GoalRepository {
 
     Optional<Goal> findById(Long goalId);
 
-    List<DayGoalsResponse.GoalListPerCategoryResponse> findDayGoals(UUID memberId, LocalDate date);
+    List<GoalListPerCategoryResponse> findDayGoals(UUID memberId, LocalDate date);
 
     List<Goal> findTotalGoals(UUID memberId, Long categoryId);
